refactor(markers): promote markers to a standalone src/markers domain
The marker model, factory, and sidecar store served both the player
(render, edit, navigate) and the recording pipeline (live marker
drafts at capture time), but lived inside src/player/, giving the
recording feature an edge into the player. They are now a top-level
src/markers/ domain that player, recording, actions, and UI all depend
on. Pure path move plus import rewrites; no behavior change.
