+++ tests/integration/AudioPlayer.regression.test.ts
@@ -21,7 +21,7 @@ import { App, Modal } from 'obsidian';
 import { AudioPlayer } from 'src/player/AudioPlayer';
 import { WaveformPeakCache, type AudioDecoder } from 'src/player/WaveformData';
 import type { AudioPlayerRegistry } from 'src/player/AudioPlayerRegistry';
-import type { MarkerStore } from 'src/player/markers/MarkerStore';
+import type { MarkerStore } from 'src/markers/MarkerStore';
 import type { ResolvedPlayerSettings } from 'src/player/playerSettings';
 import type { TFile } from 'obsidian';
