+++ tests/unit/markerFactory.test.ts
@@ -8,8 +8,8 @@
 import {
 	defaultMarkerLabel,
 	generateMarkerId,
-} from 'src/player/markers/markerFactory';
-import { MARKER_KIND, type MarkerKind } from 'src/player/markers/markerModel';
+} from 'src/markers/markerFactory';
+import { MARKER_KIND, type MarkerKind } from 'src/markers/markerModel';
 
 const kinds = (list: MarkerKind[]): { kind: MarkerKind }[] =>
 	list.map((kind) => ({ kind }));
